python-emr/requirements.txt
attrs==19.3.0
aws-cdk.aws-cloudwatch==1.38.0
aws-cdk.aws-ec2==1.38.0
aws-cdk.aws-emr==1.38.0
aws-cdk.aws-events==1.38.0
aws-cdk.aws-iam==1.38.0
aws-cdk.aws-kms==1.38.0
aws-cdk.aws-logs==1.38.0
aws-cdk.aws-s3==1.38.0
aws-cdk.aws-ssm==1.38.0
aws-cdk.cloud-assembly-schema==1.38.0
aws-cdk.core==1.38.0
aws-cdk.cx-api==1.38.0
aws-cdk.region-info==1.38.0
cattrs==1.0.0
constructs==3.0.3
jsii==1.5.0
publication==0.0.3
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.2
